+++ package.json
@@ -34,7 +34,6 @@
   },
   "devDependencies": {
     "@types/lodash": "^4.14.191",
-    "@types/numeral": "^2.0.2",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "@types/styled-components": "^5.1.26",
